requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv>=1.0.0

# Azure Neural TTS
azure-cognitiveservices-speech>=1.45.0

# Basic utilities
requests>=2.31.0
